refactor: rename the shared typography constants
