+++ src/ui/help-sections/about.tsx
@@ -1,4 +1,4 @@
-import { HEADING, PROSE } from "@/ui/help-sections/shared";
+import { HEADING, PROSE } from "@/ui/typography";
 
 // -- About --------------------------------------------------------------------
