+++ src/ui/help-sections/groups-extras.tsx
@@ -1,4 +1,4 @@
-import { HEADING, INLINE_CODE, PROSE } from "@/ui/help-sections/shared";
+import { HEADING, INLINE_CODE, PROSE } from "@/ui/typography";
 import { MOD_KEY } from "@/utils/platform";
 
 // -- Linked group extras ------------------------------------------------------
