+++ src/ui/help-sections/preview.tsx
@@ -1,4 +1,4 @@
-import { PROSE } from "@/ui/help-sections/shared";
+import { PROSE } from "@/ui/typography";
 
 // -- Preview ------------------------------------------------------------------
